requirements.txt
Flask==3.0.0
tensorflow==2.15.0
Pillow==10.1.0
numpy==1.26.2
mysql-connector-python==8.2.0
bcrypt>=4.1.2
gunicorn
